pyproject.toml
[tool.poetry]
name = "vivoback"
version = "0.1.0"
description = ""
authors = ["João M <joaomarcelo1015ba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
